refactor: reorganize the prediction module structure and remove obsolete code
