--- core/forecast/domain/repository.py
@@ -1,14 +1,0 @@
-from abc import ABC, abstractmethod
-from typing import List, Dict
-
-class MachineLearningRepository(ABC):
-    @abstractmethod
-    def getCoords(self) -> List[Dict[str, float]]:
-        pass
-
-    def getWeatherByCoord(self, lat: float, lon: float):
-        pass
-
-    @abstractmethod
-    def forecast(self, lat: float, lon: float, flood: int, date: str, probability: float):
-        pass
