+++ core/forecast/presentation/serializers/forecast.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
-from core.forecast.infra.models import Forecast
+from core.forecast.models import Forecast
 
 class ForecastSerializer(serializers.ModelSerializer):
     class Meta:
         model = Forecast
-        fields = ['latitude', 'longitude', 'date', 'flood', 'probability']
+        fields = "__all__"
