--- core/forecast/presentation/tasks/ForecastTasks.py
@@ -1,8 +1,0 @@
-from celery import shared_task
-from core.forecast.infra.repository import ForecastRepoImpl
-from core.forecast.app.services import floodingPredict
-
-@shared_task
-def forecast():
-    repo = ForecastRepoImpl()
-    floodingPredict(repo)
