+++ core/forecast/presentation/urls.py
@@ -1,6 +1,5 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from core.forecast.presentation.views.ForecastApiView import PredictView
 from core.forecast.presentation.views.ForecastViewSet import ForecastViewSet
 
 router = DefaultRouter()
@@ -8,5 +7,4 @@ router.register(r"foresee", ForecastViewSet)
 
 urlpatterns = [
     path('', include(router.urls)),
-    path("forecasts/", PredictView.as_view(), name="forecasts"),
 ]
