+++ core/forecast/presentation/views/ForecastViewSet.py
@@ -1,8 +1,7 @@
 from rest_framework.viewsets import ModelViewSet
-from core.forecast.infra.models import Forecast
-from core.forecast.presentation.serializers.ForecastSerializer import ForecastSerializer
+from core.forecast.models import Forecast
+from core.forecast.presentation.serializers.forecast import ForecastSerializer
 
 class ForecastViewSet(ModelViewSet):
     queryset = Forecast.objects.all()
     serializer_class = ForecastSerializer
-    # Deixa o Forecast livre para um CRUD manual
